backend/requirements-updated.txt
# Backend Dependencies - Community App (Updated for Auth0)

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication & Security (Auth0)
python-jose[cryptography]==3.3.0
httpx==0.25.2
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Machine Learning
torch==2.1.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
